Mis-Pruebas/ecomerce/product-service/src/main/java/com/ecommerce/productservice/service/impl/ProductServiceImpl.java
package com.ecommerce.productservice.service.impl;

import com.ecommerce.productservice.entity.Product;
import com.ecommerce.productservice.repository.ProductRepository;
import com.ecommerce.productservice.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;
@Service
public class ProductServiceImpl implements ProductService {
    @Autowired
    private ProductRepository productRepository;
    @Override
    public Optional<Product> buscarProductoPorId(Long idProducto) {
        return productRepository.findById(idProducto);
    }
}
